requirements.txt
Flask==3.1.0
g4f==0.3.2.7
python-dotenv==1.0.1
themoviedb==1.0.2
